+++ src/components/channels/index.js
@@ -1,1 +1,1 @@
-export ChannelsSubscriber from './ChannelsSubscriber';
+export { default as ChannelsSubscriber } from './ChannelsSubscriber';